+++ user-edit.php
@@ -8,7 +8,10 @@ $placeholder = require 'includes/init_page.php';
 $id_user = (int)$_GET['id_user'];
 
 if (isset($_POST['save'])) {
-    $user = new Models\User(
+    if(DEMO) {
+        $placeholder['message']['error'] = 'In der Demo Version dürfen die Teilnehmer Daten nicht geändert werden!';
+    } else {
+        $user = new Models\User(
         $id_user,
         include 'filters/post_username.php',
         include 'filters/post_name.php',
@@ -28,20 +31,29 @@ if (isset($_POST['save'])) {
         return;
     } else
         $placeholder['message']['error'] = 'Die Teilnehmer Daten konnten nicht geändert werden!';
-} elseif (isset($_POST['delete'])) {
-    if(Tables\Users::delete($database_pdo, $id_user)) {
-        header('location: user.php');
-        return;
     }
+} elseif (isset($_POST['delete'])) {
+    
+    if(DEMO) {
+         $placeholder['message']['error'] = 'In der Demo Version darf der Teilnehmer nicht gelöscht werden!';
+     } else {
+        if(Tables\Users::delete($database_pdo, $id_user)) {
+            header('location: user.php');
+            return;
+        }
+     }
 } elseif(isset($_POST['password_save']) && !empty($_POST['password'])) {
-
-    if($_POST['password'] == $_POST['password_repeat'])
-        if(Tables\Users::update_password($database_pdo, $id_user, $_POST['password']))
-            $placeholder['message']['success'] = 'Dein Passwort wurde geändert.';
+    if(DEMO) {
+        $placeholder['message']['error'] = 'In der Demo Version darf das Passwort nicht geändert werden!';
+    } else {
+        if($_POST['password'] == $_POST['password_repeat'])
+            if(Tables\Users::update_password($database_pdo, $id_user, $_POST['password']))
+                $placeholder['message']['success'] = 'Dein Passwort wurde geändert.';
+            else
+                $placeholder['message']['error'] = 'Dein Passwort konnte nicht geändert werden!';
         else
-            $placeholder['message']['error'] = 'Dein Passwort konnte nicht geändert werden!';
-    else
-        $placeholder['message']['error'] = 'Passwörter stimmen nicht überein!';
+            $placeholder['message']['error'] = 'Passwörter stimmen nicht überein!';  
+    }
 }
 
 $placeholder['user'] = Tables\Users::select_user($database_pdo, $id_user);
